# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TCG Tools",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/torneios": {
      "get": {
        "tags": [
          "torneios"
        ],
        "summary": "List Torneios",
        "operationId": "list_torneios_api_v1_torneios_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "torneios"
        ],
        "summary": "Create Torneio",
        "operationId": "create_torneio_api_v1_torneios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TorneioCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/torneios/{event_id}": {
      "get": {
        "tags": [
          "torneios"
        ],
        "summary": "Get Torneio",
        "operationId": "get_torneio_api_v1_torneios__event_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "torneios"
        ],
        "summary": "Update Torneio",
        "operationId": "update_torneio_api_v1_torneios__event_id__patch",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TorneioUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/torneios/{event_id}/jogadores": {
      "post": {
        "tags": [
          "torneios"
        ],
        "summary": "Add Jogador",
        "operationId": "add_jogador_api_v1_torneios__event_id__jogadores_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JogadorCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/torneios/{event_id}/jogadores/{player_id}": {
      "delete": {
        "tags": [
          "torneios"
        ],
        "summary": "Remove Jogador",
        "operationId": "remove_jogador_api_v1_torneios__event_id__jogadores__player_id__delete",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          },
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/torneios/{event_id}/iniciar": {
      "post": {
        "tags": [
          "torneios"
        ],
        "summary": "Iniciar",
        "operationId": "iniciar_api_v1_torneios__event_id__iniciar_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/torneios/{event_id}/rodadas": {
      "get": {
        "tags": [
          "torneios"
        ],
        "summary": "List Rodadas",
        "operationId": "list_rodadas_api_v1_torneios__event_id__rodadas_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/torneios/{event_id}/rodadas/{number}": {
      "get": {
        "tags": [
          "torneios"
        ],
        "summary": "Get Rodada",
        "operationId": "get_rodada_api_v1_torneios__event_id__rodadas__number__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          },
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/torneios/{event_id}/matches/{match_id}": {
      "patch": {
        "tags": [
          "torneios"
        ],
        "summary": "Update Match",
        "operationId": "update_match_api_v1_torneios__event_id__matches__match_id__patch",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          },
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Match Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/torneios/{event_id}/jogadores/{player_id}/drop": {
      "post": {
        "tags": [
          "torneios"
        ],
        "summary": "Drop Jogador",
        "operationId": "drop_jogador_api_v1_torneios__event_id__jogadores__player_id__drop_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          },
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DropRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/torneios/{event_id}/avancar": {
      "post": {
        "tags": [
          "torneios"
        ],
        "summary": "Avancar",
        "description": "Conclui a rodada ativa (sem iniciar a pr\u00f3xima).",
        "operationId": "avancar_api_v1_torneios__event_id__avancar_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/torneios/{event_id}/iniciar-proxima-rodada": {
      "post": {
        "tags": [
          "torneios"
        ],
        "summary": "Iniciar Proxima Rodada",
        "operationId": "iniciar_proxima_rodada_api_v1_torneios__event_id__iniciar_proxima_rodada_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/torneios/{event_id}/rodadas/reabrir": {
      "post": {
        "tags": [
          "torneios"
        ],
        "summary": "Reabrir Rodada",
        "description": "Reabre a \u00faltima rodada conclu\u00edda (ou a anterior \u00e0 rodada ativa) para corre\u00e7\u00f5es.",
        "operationId": "reabrir_rodada_api_v1_torneios__event_id__rodadas_reabrir_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          },
          {
            "name": "number",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/torneios/{event_id}/finalizar": {
      "post": {
        "tags": [
          "torneios"
        ],
        "summary": "Finalizar",
        "operationId": "finalizar_api_v1_torneios__event_id__finalizar_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/torneios/{event_id}/premiacao": {
      "get": {
        "tags": [
          "torneios"
        ],
        "summary": "Get Premiacao",
        "operationId": "get_premiacao_api_v1_torneios__event_id__premiacao_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/torneios/{event_id}/classificacao": {
      "get": {
        "tags": [
          "torneios"
        ],
        "summary": "Get Classificacao",
        "operationId": "get_classificacao_api_v1_torneios__event_id__classificacao_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "torneios"
        ],
        "summary": "Patch Classificacao",
        "operationId": "patch_classificacao_api_v1_torneios__event_id__classificacao_patch",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassificacaoPatch"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/torneios/{event_id}/export-log": {
      "post": {
        "tags": [
          "torneios"
        ],
        "summary": "Export Log",
        "operationId": "export_log_api_v1_torneios__event_id__export_log_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClassificacaoPatch": {
        "properties": {
          "updates": {
            "items": {
              "$ref": "#/components/schemas/DecklistUpdate"
            },
            "type": "array",
            "title": "Updates"
          }
        },
        "type": "object",
        "required": [
          "updates"
        ],
        "title": "ClassificacaoPatch"
      },
      "DecklistUpdate": {
        "properties": {
          "player_id": {
            "type": "integer",
            "title": "Player Id"
          },
          "decklist": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decklist"
          }
        },
        "type": "object",
        "required": [
          "player_id"
        ],
        "title": "DecklistUpdate"
      },
      "DropRequest": {
        "properties": {
          "mid_round": {
            "type": "boolean",
            "title": "Mid Round",
            "default": false
          }
        },
        "type": "object",
        "title": "DropRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JogadorCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "JogadorCreate"
      },
      "MatchUpdate": {
        "properties": {
          "score_p1": {
            "type": "integer",
            "title": "Score P1"
          },
          "score_p2": {
            "type": "integer",
            "title": "Score P2"
          }
        },
        "type": "object",
        "required": [
          "score_p1",
          "score_p2"
        ],
        "title": "MatchUpdate"
      },
...
      "TorneioCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "event_date": {
            "type": "string",
            "format": "date",
            "title": "Event Date"
          },
          "format": {
            "type": "string",
            "title": "Format"
          },
          "max_rounds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Rounds"
          },
          "entry_fee": {
            "type": "number",
            "title": "Entry Fee",
            "default": 0
          },
          "best_of": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Best Of",
            "default": 3
          },
          "premiacao_preset_id": {
            "type": "string",
            "title": "Premiacao Preset Id",
            "default": "standard"
          }
        },
        "type": "object",
        "required": [
          "name",
          "event_date",
          "format"
        ],
        "title": "TorneioCreateRequest"
      },
      "TorneioUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "event_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Date"
          },
          "entry_fee": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Fee"
          },
          "best_of": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Of"
          },
          "max_rounds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Rounds"
          }
        },
        "type": "object",
        "title": "TorneioUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}